altos: Use TeleMetrum v1.9 board for MegaDongle experiments
This configures the MegaDongle v0.1 directory to use the pin
assignments in TeleMetrum v1.9 boards.

+++ src/kernel/ao_rssi.c
@@ -33,7 +33,7 @@ ao_rssi(void)
 }
 
 void
-ao_rssi_set(int rssi_value)
+ao_rssi_set(int16_t rssi_value)
 {
 	if (rssi_value > 0)
 		rssi_value = 0;
